feat: expose lock & add timeout on push job so that queue cant be stuck
BREAKING CHANGE: earlier it was not needed to ensure that we operate on
the taken lock, and it was possible for it to expire during long jobs
and another process to take control of the job creating multiple
parallel threads. To avoid that `onJobComplete` functions exposes `lock`
as a property, which allows us to preserve back-compatibility. However,
it must be noted that user must take care of lock extension during
lengthy tasks and that they will fail with a timeout error if result
is not resolved within default lock timeout multiplied by 2 or by
expected custom timeout provided to `push` function.